Schema version 0.0.4 implemented
A schema that supports list-views is now implemented.  Required changing template (slightly) and adding a list-view template "view_collection_list.html".  Code-wise, I added a "login_required" very similar to the basic one implemented in Flask.  But it is part of the Admin object.  I changed `_schema_transform` to support the leading caret and asterisk.  These are reflected in extra key-values in fields.  A bit awkward, but good redundancy was added with the mongodb `_id` added to every field.  It is mostly unused.  I also changed the `view_collection(coll)` method of Admin to use the new schema to service the list-view template.